+++ packages/deprecated/stylus/package.js
@@ -1,0 +1,8 @@
+Package.describe({
+  summary: 'DEPRECATED - Expressive, dynamic, robust CSS',
+  version: "2.513.14"
+});
+
+Package.onUse(function (api) {
+  api.addFiles('deprecation_notice.js');
+});